% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nbins.R
\name{find_nbins}
\alias{find_nbins}
\title{Compute a reasonable default number of bins.}
\usage{
find_nbins(x, rule = c("FD", "RR", "Scott", "sqrt", "Sturges"), na_rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{rule}{a character string matching to a rule to use to determine
number of bins.}

\item{na_rm}{logical whether to remove (\code{TRUE}) missing values or not
(\code{FALSE})}
}
\value{
an integer with reasonable number of bins.
}
\description{
Compute a reasonable default number of bins based on the data and different
types of rules.
}
\examples{
set.seed(42L)
find_nbins(rnorm(1000))
find_nbins(rnorm(1000), "Scott")
find_nbins(rnorm(1000), "sqrt")
}
\seealso{
The following
\href{https://en.wikipedia.org/wiki/Histogram#Number_of_bins_and_width}{wiki}
for a discussion on different "rules-of-thumb". Additionally, the
\href{https://github.com/hadley/bigvis}{bigvis} package and sources therein.
}
