% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check if a \code{R} object is indeed empty}
\usage{
is_empty(x, trim = TRUE)
}
\arguments{
\item{x}{a \code{R} object}

\item{trim}{a logical scalar which controls to trim leading and trailing
whitespaces with a character input.}
}
\value{
a logical vector.
}
\description{
A helper function to check whether a \code{R} object is indeed empty mostly
used internally.
}
\examples{

# Empty strings
is_empty(" ")               # TRUE
is_empty(character(0))      # TRUE
is_empty(" ", trim = FALSE) # FALSE

# Case where it returns TRUE
is_empty(0)
is_empty(numeric(0))
}
