% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse_cg}
\alias{sparse_cg}
\title{Conjugate Gradient solver for Sparse Matrices.}
\usage{
sparse_cg(A, b, x0, maxiter = 10000L, tol = 1e-06)
}
\arguments{
\item{A}{Sparse double matrix of class \linkS4class{dgCMatrix}.}

\item{b}{Numeric vector with length equal to number of rows of \code{A}.}

\item{x0}{Numeric vector with same length as \code{b}. If no initial guess
is known a vector of zeros with length of \code{b} can be passed.}

\item{maxiter}{Integer scalar that sets the maximum number of iterations
allowed by the solver. Default is 10,000.}

\item{tol}{Sets the tolerance threshold (upper bound to the relative
residual error) used by the stopping criteria. Default is \code{1e-6}.}
}
\value{
\code{sparse_cg} returns a list with several components:
\item{coefficients}{a vector with the estimated coefficients}
\item{itr}{a scalar denoting the number of iterations at convergence}
\item{err}{a scalar denoting the relative error at convergance}
}
\description{
Using the iterative Conjugate Gradient Solver from the Eigen C++ library
it solves a linear system \code{A * x = b}. Currently, it uses a diagonal
preconditioner.
}
\seealso{
See \href{http://www.eigen.tuxfamily.org/}{Eigen Library} for more
documentation on solver used and other available solvers.
}
