% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_lm.R
\name{sparse_lm}
\alias{sparse_lm}
\title{Fit Linear Model using sparse matrices}
\usage{
sparse_lm(X, y, x0, maxiter = 4L * ncol(X), tol = .Machine$double.eps)
}
\arguments{
\item{X}{a sparse matrix of the class \code{\link[Matrix]{dgCMatrix-class}}.}

\item{y}{a numeric vector with response variable.}

\item{x0}{a numeric vector of length \code{ncol(X)} with an initial guess.
If it's not provided it will default to a zero-vector.}

\item{maxiter}{an integer scalar with the maximum number of iterations.
A good choice is usually \code{4 * ncol(X)}.}

\item{tol}{a numeric scalar with the relative error tolerance.}
}
\value{
A list with the elements

\item{coefficients}{The estimated coefficients from solver.}
\item{itr}{A scalar denoting the number of iterations.}
\item{error}{A scalar denoting the the relative error at the last iteration.}
\item{fitted.values}{A vector with the fitted values.}
\item{residuals}{The vector with the residuals.}
}
\description{
\code{sparse_lm} uses sparse matrices and a iterative solver to fit linear models.
The sparse iterative solvers are from the \code{Eigen} C++ library.
Currently, only the Conjugate Gradient algorithm has been implemented with
a diagonal preconditioner.
}
\seealso{
See
\href{http://eigen.tuxfamily.org}{Eigen C++ library} for documentation on
\code{Eigen} library and
\href{https://en.wikipedia.org/wiki/Conjugate_gradient_method}{Conjugate Gradient}
for more information about this method.
}
\keyword{models}
\keyword{regression}
