% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_list.R
\name{str_list}
\alias{str_list}
\title{List items}
\usage{
str_list(string, sep = ", ", last = " and ")
}
\arguments{
\item{string}{an object coercible to a character vector.}

\item{sep}{a string to seperate elements except last.}

\item{last}{a string to seperate last element.}
}
\value{
a character string.
}
\description{
\code{str_list()} wraps \code{\link[stringr:str_c]{stringr::str_c()}} in order to turn a list of character
strings into a grammatically correct list without an
\href{https://en.wikipedia.org/wiki/Serial_comma}{Oxford Comma} by default.
Sue me.
}
\examples{
shopping_list <- c("bread", "milk", "eggs")
str_list(shopping_list)
}
