% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_msg.R
\name{str_msg}
\alias{str_msg}
\title{Message friendly string interpolation.}
\usage{
str_msg(..., env = parent.frame())
}
\arguments{
\item{...}{one or more character vectors which are passed to
\code{\link[stringr:str_c]{stringr::str_c()}}. Multiple strings are automatically joined
with a space as seperator.}

\item{env}{the environment in which to evaluate the expressions.}
}
\value{
an interpolated character string.
}
\description{
This is a wrapper function around \code{\link[stringr:str_interp]{stringr::str_interp()}} and
\code{\link[stringr:str_c]{stringr::str_c()}}. It aids in creating useful messages which can
be passed to \code{\link[=base:warning]{base:warning()}} or \link[base:stop]{base::stop}, for example, or other
logging functions.
}
\examples{
user_name <- "World"
str_msg("Hello ${user_name}!", "Welcome to tidyverse.")
}
\seealso{
\code{\link[stringr:str_interp]{stringr::str_interp()}} and \code{\link[stringr:str_c]{stringr::str_c()}} which this
function wraps for further documenation or examples on string
interpolation or the joining of strings.
}
