% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_padding.R
\name{str_padding}
\alias{str_padding}
\alias{str_lpad}
\alias{str_rpad}
\alias{str_bpad}
\title{Pad a string (with better defaults).}
\usage{
str_lpad(string, width, pad = " ")

str_rpad(string, width, pad = " ")

str_bpad(string, width, pad = " ")
}
\arguments{
\item{string}{A character vector.}

\item{width}{Minimum width of padded strings.}

\item{pad}{Single padding character (default is a space).}
}
\value{
a padded character vector.
}
\description{
These functions just wrap around \code{\link[stringr:str_pad]{stringr::str_pad()}} with some better
defaults (imho). If \code{width} is missing it defaults to length of the
\code{string + 1} for \code{str_lpad()} and \code{str_rpad()} and \code{+ 2} for \code{str_bpad()}.
}
\examples{
rbind(
  str_lpad("hadley", 30, "_"),
  str_rpad("hadley", 30, "-"),
  str_bpad("hadley", 30, "'")
)
}
\seealso{
\code{\link[stringr:str_pad]{stringr::str_pad()}} for more examples.
}
