% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstevR.R
\docType{package}
\name{tstevR-package}
\alias{tstevR}
\alias{tstevR-package}
\title{tstevR: Someone's Collection of R functions}
\description{
A non-impressive collection of functions gathered over time. That
    or a good excuse to start dabbling in R programming and package building.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tstev/tstevR}
  \item Report bugs at \url{https://github.com/tstev/tstevR/issues}
}

}
\author{
\strong{Maintainer}: Tomas Stevens \email{t.stevens90@gmail.com} [copyright holder]

}
\keyword{internal}
