% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstevR.R
\docType{package}
\name{tstevR}
\alias{tstevR}
\alias{tstevR-package}
\title{\code{tstevR} package}
\description{
Someone's Collection of R functions
}
\details{
A non-impressive collection of functions gathered over time. That or a good
excuse to start dabbling in R programming and package building.
}
